move modelinfo  and modelinfo conversion to proper places
